# internal/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "BinRag 企业级文档知识库问答系统 API。所有接口需 `Authorization: Bearer \u003cAPI Key\u003e`（除 Swagger 文档页本身）。统一响应格式 `{code, message, data}`，code=0 表示成功，非 0 为业务码（400/401/404/500）。",
        "title": "BinRag API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/auth/exchange": {
            "post": {
                "description": "用回调后拿到的一次性 ticket 换取会话 JWT；ticket 仅可成功消费一次",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "换取会话 JWT",
                "parameters": [
                    {
                        "description": "一次性 ticket",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.exchangeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "token": {
                                                    "type": "string"
                                                }
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/auth/github/login": {
            "get": {
                "description": "GitHub 使用 OAuth2 适配（非 OIDC），302 到 GitHub 授权页",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "发起 GitHub 登录",
                "responses": {
                    "302": {
                        "description": "Found"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/me": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "返回当前凭据身份：apikey → {kind,is_bootstrap}；oidc → {kind,user_id,provider,name}",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "当前身份",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/oidc/{provider}/callback": {
            "get": {
                "description": "OIDC 授权码回调；成功后 302 到 /login?ticket=xxx（一次性，前端换取 JWT）。校验失败不创建会话",
                "tags": [
                    "认证"
                ],
                "summary": "OIDC 回调",
                "parameters": [
                    {
                        "type": "string",
                        "description": "OIDC provider 名",
                        "name": "provider",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Found"
                    }
                }
            }
        },
        "/api/v1/auth/oidc/{provider}/login": {
            "get": {
                "description": "按 provider 名发起 OIDC 授权码登录，302 重定向到对应 issuer 授权页",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "发起 OIDC 登录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "OIDC provider 名（见 /auth/providers）",
                        "name": "provider",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Found"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/chat": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "接收问题与可选的会话、知识库范围，走 RAG 编排返回回答与引用来源。默认返回 JSON；请求带 Accept: text/event-stream 或 query stream=1 时返回 SSE 流式（事件序列：thinking×N 思考链路 → sources 引用来源 → chunk×N 文本增量 → done 正常结束 / error 出错终止；思考链路开启时才有 thinking 事件）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "text/event-stream"
                ],
                "tags": [
                    "问答"
                ],
                "summary": "问答",
                "parameters": [
                    {
                        "description": "问答请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.chatRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/rag.RAGResult"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/chat/enhancements": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "返回可用增强能力。当前能力：web_search（联网搜索）；available 表示后端是否已配置就绪（配置了 web_search.api_key）",
                "tags": [
                    "问答"
                ],
                "summary": "增强能力列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "api.chatRequest": {
            "type": "object",
            "required": [
                "question",
                "session_id"
            ],
            "properties": {
                "enhanced": {
                    "description": "增强模式（function calling 工具，如 web_search）",
                    "type": "boolean"
                },
                "kb_id": {
                    "description": "知识库范围，空表示不限定",
                    "type": "string"
                },
                "question": {
                    "type": "string"
                },
                "session_id": {
                    "type": "string"
                },
                "strategy": {
                    "description": "单次请求策略覆盖",
                    "allOf": [
                        {
                            "$ref": "#/definitions/config.StrategyConfig"
                        }
                    ]
                }
            }
        },
...
        "api.exchangeRequest": {
            "type": "object",
            "required": [
                "ticket"
            ],
            "properties": {
                "ticket": {
                    "type": "string"
                }
            }
        },
...
        "config.StrategyConfig": {
            "type": "object",
            "properties": {
                "data_sources": {
                    "description": "DataSources 允许的数据源列表；空=未设置（合并后默认仅 vector_store）",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "decomposition": {
                    "type": "string"
                },
                "fusion": {
                    "type": "string"
                },
                "hyde": {
                    "type": "string"
                },
                "query": {
                    "type": "string"
                },
                "routing": {
                    "type": "string"
                },
                "step_back": {
                    "type": "string"
                },
                "thinking": {
                    "description": "off / on，空=继承",
                    "type": "string"
                }
            }
        },
...
        "rag.RAGResult": {
            "type": "object",
            "properties": {
                "answer": {
                    "type": "string"
                },
                "sources": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rag.Source"
                    }
                },
                "thinking": {
                    "description": "思考链路（F8，关闭时省略）",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rag.ThinkingStep"
                    }
                }
            }
        },
        "rag.Source": {
            "type": "object",
            "properties": {
                "filename": {
                    "description": "来源文件名（来自元数据）",
                    "type": "string"
                },
                "heading": {
                    "description": "标题上下文",
                    "type": "string"
                },
                "id": {
                    "description": "检索片段 ID",
                    "type": "string"
                },
                "score": {
                    "description": "检索分数",
                    "type": "number"
                },
                "source_type": {
                    "description": "来源类型（vector_store / web_search 等）",
                    "type": "string"
                }
            }
        },
        "rag.ThinkingStep": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "环节结构化数据"
                },
                "elapsed_ms": {
                    "description": "环节耗时",
                    "type": "integer"
                },
                "label": {
                    "description": "前端展示标题",
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/rag.ThinkingStepType"
                }
            }
        },
        "rag.ThinkingStepType": {
            "type": "string",
            "enum": [
                "routing",
                "query_rewrite",
                "multi_query",
                "retrieval",
                "rerank",
                "chunks",
                "decompose",
                "step_back",
                "hyde",
                "tool"
            ],
            "x-enum-comments": {
                "StepChunks": "最终目标 chunks",
                "StepDecompose": "分解判定 + 子问题列表",
                "StepHyDE": "假设文档",
                "StepMultiQuery": "多查询变体生成",
                "StepRerank": "重排序前后对比",
                "StepRetrieval": "检索（单路或多路）",
                "StepRewrite": "单查询改写",
                "StepRouting": "路由判定",
                "StepStepBack": "回退问题",
                "StepTool": "工具调用（增强模式 function calling）"
            },
            "x-enum-varnames": [
                "StepRouting",
                "StepRewrite",
                "StepMultiQuery",
                "StepRetrieval",
                "StepRerank",
                "StepChunks",
                "StepDecompose",
                "StepStepBack",
                "StepHyDE",
                "StepTool"
            ]
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}